Renamees eventqueue to queue, and move queue and batcher under events
