+++ events/batcher/batcher.go
