+++ events/batcher/batcher_test.go
