+++ events/batcher/batcher_unit_test.go
